[ART-2056] Rename `find-cve-flaws` to `attach-cve-flaws`
