+++ elliottlib/cli/__main__.py
@@ -58,7 +58,7 @@ from elliottlib.cli.verify_cvp_cli import verify_cvp_cli
 from elliottlib.cli.advisory_drop_cli import advisory_drop_cli
 from elliottlib.cli.verify_attached_operators_cli import verify_attached_operators_cli
 from elliottlib.cli.verify_attached_bugs_cli import verify_attached_bugs_cli
-from elliottlib.cli.find_cve_flaws_cli import find_cve_flaws_cli
+from elliottlib.cli.attach_cve_flaws_cli import attach_cve_flaws_cli
 
 # 3rd party
 import bugzilla
@@ -694,7 +694,7 @@ cli.add_command(verify_cvp_cli)
 cli.add_command(advisory_drop_cli)
 cli.add_command(verify_attached_operators_cli)
 cli.add_command(verify_attached_bugs_cli)
-cli.add_command(find_cve_flaws_cli)
+cli.add_command(attach_cve_flaws_cli)
 
 
 # -----------------------------------------------------------------------------
